backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
greenlet>=3.0.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
python-telegram-bot==21.9
apscheduler==3.10.4
sentry-sdk[fastapi]==2.19.2
httpx==0.28.1

# Dev / test
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite>=0.20.0
